implement 'mut' for struct/tuple elements
Changing the value of a struct-field or tuple-element is now only possible if that field is declared as mutable (field type ends with ' mut').
The 'mut' at the end of the field type only influences if the variable can be modified, otherwise it is ignored.

Removed implicit type-casting of tuple elements at runtime.