introduce type fields and pseudo fields
type fields are fields that are directly linked to a type, "pseudo fields" are special type fields, they contain procedures that take a value of that type as last argument. Pseudo fields cannot be accessed directly, instead the procedures are implicitly evaluated when the field is accessed on a value of that type.